requirements.txt
fastapi==0.111.0
uvicorn==0.30.5
pydantic==2.8.2
python-dotenv==1.0.1
jinja2==3.1.4
numpy==1.26.4
scikit-learn==1.5.1
prometheus-client==0.20.0
requests==2.32.3
faiss-cpu==1.8.0.post1
PyJWT==2.9.0
scipy==1.13.1
rank-bm25==0.2.2
rapidfuzz==3.9.6
PyPDF2==3.0.1
python-docx==1.1.2
